requirements.txt
asn1crypto==1.3.0
astroid==2.3.3
autopep8==1.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.13.0
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wrapt==1.12.1
